+++ app/test/test_helpers/session_test_helper.rb
@@ -9,8 +9,10 @@ module SessionTestHelper
     ActionDispatch::TestRequest.create.cookie_jar.tap do |cookie_jar|
       cookie_jar.encrypted[:session_id] = Current.session.id
       cookie_jar.encrypted[:session_tenant_id] = user.tenant_id
+      cookie_jar.encrypted[:session_user_uuid_id] = user.uuid_id
       cookies["session_id"] = cookie_jar[:session_id]
       cookies["session_tenant_id"] = cookie_jar[:session_tenant_id]
+      cookies["session_user_uuid_id"] = cookie_jar[:session_user_uuid_id]
     end
   end
 
@@ -19,6 +21,7 @@ module SessionTestHelper
     Current.user = nil
     cookies.delete("session_id")
     cookies.delete("session_tenant_id")
+    cookies.delete("session_user_uuid_id")
   end
 end